+++ .eslintrc.js
@@ -7,20 +7,12 @@ module.exports = {
     extends: [
         'eslint:recommended',
         'plugin:@typescript-eslint/recommended',
-        'plugin:md/recommended',
     ],
     rules: {
         "no-unused-vars": "off",
         "@typescript-eslint/no-unused-vars": "warn",
         "@typescript-eslint/explicit-module-boundary-types": "off",
         "@typescript-eslint/indent": ["error"],
-        'md/remark': [
-            'error',
-            {
-                // This object corresponds to object you would export in .remarkrc file
-                plugins: ['preset-lint-markdown-style-guide', 'lint', ['lint-list-item-spacing', false]],
-            },
-        ],
     },
     overrides: [
         {
@@ -37,27 +29,5 @@ module.exports = {
                 "no-undef": "off"
             }
         },
-        {
-            files: ['*.md'],
-            parser: 'markdown-eslint-parser',
-        },
-        {
-            files: ['CHANGELOG.md'],
-            rules: {
-                'md/remark': [
-                    'error',
-                    {
-                        // This object corresponds to object you would export in .remarkrc file
-                        plugins: [
-                            'preset-lint-markdown-style-guide',
-                            'lint',
-                            ['lint-list-item-spacing', false],
-                            ['lint-no-shortcut-reference-link', false],
-                            ['lint-no-duplicate-headings', false],
-                        ],
-                    },
-                ],
-            }
-        }
     ]
 };